+++ db/migrate/20171024075514_customer_vault_contact_type_default_value.rb
@@ -1,0 +1,5 @@
+class CustomerVaultContactTypeDefaultValue < ActiveRecord::Migration[5.0]
+  def change
+    Dorsale::CustomerVault::Event.where(contact_type: nil).update_all(contact_type: "contact")
+  end
+end